requirements.txt
transformers
accelerate
pillow
pandas
torch
einops
numpy<2.0.0
